refactor(config): split preflight.py into package with validators/ subdirectory
Split the 663-line preflight.py monolith into a preflight/ package with
validators grouped under validators/ subdirectory by config domain:

preflight/
├── __init__.py         - public API (run_preflight_check) + re-exports
├── models.py           - data models (CheckStatus, CheckResult, etc.)
└── validators/
    ├── __init__.py     - re-export all validators
    ├── orchestrator.py - orchestrator config validation
    ├── team.py         - team config validation
    ├── evaluator.py    - evaluator, custom metrics, metric name validation
    ├── judgment.py     - judgment config validation
    ├── prompt_builder.py - prompt builder config validation
    ├── auth.py         - auth credential validation + model ID collection
    └── workspace.py    - workspace writable check

All existing imports (from mixseek.config.preflight import ...) remain
backward-compatible via __init__.py re-exports. No logic changes.
